refactor: 예약자명을 MemberName으로 분리
